Update myfile.py
I am quite confused by this... I am still unsure how did the numbers come about. So, I do not know how I want my code to go. I know that we need to make a loop, but is there another function I can call that transcribe the A, G, C, T for me? then I call it, implement it into the loop, and then... Turns out. I figured it out. I know where the numbers are from. and now I think I have a way to do this. I will set "A" as and variable and the same with the rest. Then make a loop that every time it sees "A", it becomes i + 1. However, does that mean I need to make separate loops for all the letters?????? I could use selective, or I could use the bubble method where arrange the letters first. I will put all the "A" together and vice versa for the rest. Then, I simple make a function that counts them at the end. I think that would be the best method. However, I am unsure with where to start. But after I start it, I think I can finish it.