Update myfile.py
I forgot to update my changes for this morning. I am trying to figure out what is the best direction to take this. I know for sure I want to do a bubble instead of selection. but if A is A, then what value do i put? as it is not I???