Update myfile.py
Today is Wednesday January 13th. I started to debug problem 1 and started on problem 2. I think I am close to being done for both but there a little mistakes here and there that I cannot find. I worked from 1:30pm to 2:20pm. Then I continued from 3:00pm - 3:30pm. It does not look like I did much, but I just kept trying to fix problem 1 but every change makes it a bit more complicated. I also tried to google an explanation for problem 3 as I did not understand it. But it seems like everyone is using Java instead of Visual Studio Code.

I do not know what is wrong with problem 2, because I took a lot of inspiration from our lab 2 where we did the reverse. However, I am only transcribing the letter T and I do not know if I need a place holder for that.

I also took a look at problem 3, but I see no pattern there to make a loop for. It was similar related to pairings but I am unsure about that. I already made an office hours appointment for thursday evening, where I will be asking Ms. DeRanek to debug problems 1 and 2 with me and also ask her to extend on problem 3 so I know how to continue it.