Update myfile.py
Today, I went to office hours with Ms. DeRanek. She reminded me that = does not ==, so that helped a lot for my "ongoing" assignments. I made changes for those already. then I learned about the string and integer. I also learned that my string will never be smaller than 0, so my while loop does not work! So now, in my problem 2, I cannot tell if the T in my string/sequence is a string or and list... I shall ask. On the bring side, I debugged problem 3 all on my own! honestly something I have not done in a while and I am proud and excited. Then Willma came in with struggles of problem 4, so now I am worried.