Update myfile.py
Today is Thursday Jan 21. i spent the last two work sessions on problem 4 because I did not really understand the "hamming" and I could not figure out how to start. I have not commited in a while because I fell into a rabbit hole of solving my most recent file. However I finally did it!!!!!! It did take me a while and I spend a the last work session simply trying many different things. at first I tried a while loop to maybe say if these strings = each other then I would keep running the file. But then that was not the case because they will always equal each other and my code will keep running forever. Then I was thinking maybe if I set one dnaseq to be the "boss" so other strings are being compared to my one "boss" so that was when I came up with DNASEQ1. Then I can just do a for loop and compare the i in all of the sequence. But the catch is that whenever I see a match, I will add it to the "hamming distance" variable that I set before hand (learned that from lab 2). Then after I compare and add it to the hamming distance, I have to move on and look at the other letters. So, I made the computer move on afterwards. I am super relieved that I finally did it. I was in such a deep rabbit hole but I feel like a weight has been lifted off my coding fingers. Now I am excited to move on yay!